Add nettle-aes
This is really a proof of concept so I could see how much work it was to
add a new Nettle benchmark, but I think it does work and is sensible to
include given that DES is more than a bit crusty.  I've chosen AES128
here as I think it's more appliciable in an embedded context.

This is essentially just a copy of the DES benchmark, but swapped out
for AES.  I noticed a few things about the DES benchmark:

* It appears that the key is never set within `des_ctx`.  I think this
  is probably OK for the symmetric crypto benchmarks, but it is a bit
  odd.  I've done this for AES as well.
* There's only a single block of encryption/decryption.  This is
  probably fine for a compiler benchmark, but it seems odd if we're
  looking into benchmarking the hardware.

I haven't done much here, but I did at least go and verify that changing
the expected output caused a test failure on my workstation.

Signed-off-by: Palmer Dabbelt <palmer@dabbelt.com>